Move test.py to main.py
